Rename pretty much all references to r2c
