+++ .gitmodules
@@ -1,0 +1,3 @@
+[submodule "RoR2CommandFinder"]
+	path = RoR2CommandFinder
+	url = https://github.com/harbingerofme/RoR2CommandFinder
